wip: Refine Bag DTOs to fix economy services compilation (Phase P1) Bag DTO Improvements: + Added nested Item class to BagAddItemReq + Added nested Cost class to BagConsumeReq + Changed itemId type from Long to Integer (matches service usage) + Added ok() and error() instance methods + Added static factory methods (ok, fail) + Renamed success fields to 'succeeded' to avoid Lombok conflicts + Added @JsonProperty annotations for JSON compatibility + Added convenience constructors + Added reason/reasonType/idemKey fields Files Modified: - BagAddItemReq.java - Nested Item class, Integer types - BagAddItemResp.java - ok()/error() methods, succeeded field - BagConsumeReq.java - Nested Cost class, batch support - BagOkResp.java - ok()/error() methods, succeeded field - BagDTOs.java - ItemDelta Integer type Error Reduction: - Before: 31 errors in 5 economy services - Expected: Most Bag DTO errors resolved - Remaining: common-lib has 57 OTHER errors (Lombok issues in other DTOs) Status: ⚠️ common-lib build still failing (Lombok annotation issues in Box, Equip, Drop DTOs) ✅ Bag DTO structure matches service expectations ⏳ Next: Fix remaining common-lib Lombok errors Session Summary: + MIGRATION_SESSION_SUMMARY.md - Complete progress report   - Documentation analysis results   - DTO refactoring status   - Phase P0/P1 progress metrics   - Next steps and recommendations   - Migration strategy from C++ code See docs/MIGRATION_SESSION_SUMMARY.md for complete session report

+++ common-lib/src/main/java/org/SouthMillion/dto/bag/BagDTOs.java
@@ -27,7 +27,7 @@ public class BagDTOs {
          * Item ID
          */
         @NotNull
-        private Long itemId;
+        private Integer itemId;
 
         /**
          * Amount (positive for add, negative for consume)
@@ -46,13 +46,13 @@ public class BagDTOs {
         private Boolean bound;
 
         // Convenience constructor
-        public ItemDelta(Long itemId, Integer amount) {
+        public ItemDelta(Integer itemId, Integer amount) {
             this.itemId = itemId;
             this.amount = amount;
         }
 
         // Record-style accessors
-        public Long itemId() {
+        public Integer itemId() {
             return itemId;
         }
